避免非xml文件的干扰

+++ XmlParseAlive.py
@@ -8,10 +8,12 @@ import ConfigUtil
 def clean(start,end):
     dir = os.getcwd() + '/resources/danmu/all/'
     for file in os.listdir(dir):
-        name_id = int(file.split('-')[1].split('.')[0])
-        if start <= name_id <= end:
-            os.remove(dir + file)
-            print(file + " deleted")
+        # 避免非xml文件的干扰
+        if file.split('.')[1] == 'xml':
+            name_id = int(file.split('-')[1].split('.')[0])
+            if start <= name_id <= end:
+                os.remove(dir + file)
+                print(file + " deleted")
     print('cleaned xml finished.')
